% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_create_grid.R
\name{.get_estimate_list}
\alias{.get_estimate_list}
\title{Produces estimate list, formatted for Google Sheet, from the provided df}
\usage{
.get_estimate_list(estimate_df)
}
\arguments{
\item{estimate_df}{The tibble you want a summary of}
}
\value{
A list where each line is the summary of a row in \code{estimate_df}
}
\description{
Produces estimate list, formatted for Google Sheet, from the provided df
}
\examples{
est_info <- .get_estimate_list(estimate_df)
}
