% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_spec_file.R
\name{.parse_spec_file}
\alias{.parse_spec_file}
\title{Parse .yaml-format spec file}
\usage{
.parse_spec_file(spec_fpath = NULL)
}
\arguments{
\item{spec_fpath}{}
}
\value{
Parsed spec object
}
\description{
Parse .yaml-format spec file
}
