% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_parse.R
\name{.process_data_file}
\alias{.process_data_file}
\title{Process a raw data file, transforming it into Threadsheets format}
\usage{
.process_data_file(fpath, spec, verbose = FALSE)
}
\arguments{
\item{fpath}{The filepath to the dataset you want to process}

\item{spec}{The parsed specification data (returned by \code{.parse_spec_file()})}

\item{verbose}{Optional: Set to \code{TRUE} to print more debugging info than usual}
}
\value{
The long-format tibble where each row is an estimate derived from the data file \code{fpath}

long_df <- .process_data_file("./data/cp_membership_num/file.csv", spec, verbose = TRUE)
}
\description{
This is the main function for this file
}
