% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_create_grid.R
\name{.save_dfs}
\alias{.save_dfs}
\title{Helper function that saves the numeric and info dfs together}
\usage{
.save_dfs(num_df, info_df, output_path, grid_varname, fname_suffix)
}
\arguments{
\item{fname_prefix}{}
}
\description{
Helper function that saves the numeric and info dfs together
}
