% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_parse.R
\name{.var_in_df}
\alias{.var_in_df}
\title{Check if a variable is present in a tibble}
\usage{
.var_in_df(varname, df)
}
\arguments{
\item{varname}{}
}
\description{
We need this because of the special '@' character, which separates levels
of the MultiIndex
}
