% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_create_grid.R
\name{get_estimate_list}
\alias{get_estimate_list}
\title{Produces estimate list, formatted for Google Sheet, from the provided df}
\usage{
get_estimate_list(estimate_df)
}
\description{
Produces estimate list, formatted for Google Sheet, from the provided df
}
