% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_combine.R
\name{threader_combine}
\alias{threader_combine}
\title{Step 2: Combine Parsed Data Files}
\usage{
threader_combine(parsed_path = NULL, verbose = FALSE)
}
\arguments{
\item{parsed_path}{Path to the directory containing the parsed files generated by \code{threader_parse()}.}

\item{verbose}{Optional: If \code{TRUE}, prints debugging information}
}
\value{
Output: a giant csv file with a row for \emph{every} piece of data we have
}
\description{
Step 2: Combine Parsed Data Files
}
\examples{
combined_df <- threader_combine()
}
