% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_create_grid.R
\name{threader_create_grid}
\alias{threader_create_grid}
\title{Step 3: Construct the Grid of Estimates}
\usage{
threader_create_grid(all_entries_fpath = NULL, spec_fpath = NULL)
}
\arguments{
\item{all_entries_fpath}{The filepath to the combined .rds file generated by \code{threader_combine()}.}

\item{spec_fpath}{The filepath to the \code{.yaml}-format spec file}
}
\value{
Wide-format tibble, with one row per unit of observation and one column per unit of time
}
\description{
Step 3: Construct the Grid of Estimates
}
\examples{
grid_df <- threader_create_grid()
}
