% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_parse.R
\name{threader_parse}
\alias{threader_parse}
\title{Step 1: Parse Data Files}
\usage{
threader_parse(data_path = NULL, spec_fpath = NULL, verbose = FALSE)
}
\arguments{
\item{data_path}{Path to the directory containing the raw data you want to parse.}

\item{spec_fpath}{Filepath to the \code{.yaml}-format spec file}

\item{verbose}{Optional: If \code{TRUE}, prints additional debugging information.}
}
\value{
An environment where each key is an input file, and each value is the tibble generated from the data in that file
}
\description{
Step 1: Parse Data Files
}
\examples{
all_dfs <- threader_parse()
}
