% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_push_to_git.R
\name{threader_push_to_git}
\alias{threader_push_to_git}
\title{Push estimate grid to GitHub}
\usage{
threader_push_to_git(
  output_path = "./data/parsed/",
  spec_fpath = "./data/demo.yaml"
)
}
\value{
An environment with keys \code{num_result}, the return value from the GH API call pushing the numeric grid, and \code{info_result}, the return value from the GH API call pushing the info grid
}
\description{
Push estimate grid to GitHub
}
\examples{
gh_result <- threader_push_to_git()
}
