% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_sumstats.R
\name{threader_sumstats}
\alias{threader_sumstats}
\title{Print a list of summary statistics}
\usage{
threader_sumstats(grid_long_fpath = NULL)
}
\arguments{
\item{grid_fpath}{The filepath to the generated \emph{long-format} grid data}
}
\value{
A list with keys \code{num_obs}, \code{mean_est}, \code{num_nonnull}, and \code{num_null}.
}
\description{
Can be used as an initial diagnostic, to check that the pipeline ran successfully
}
\examples{
grid_sumstats <- threader_sumstats()
}
