% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_create_grid.R
\name{compute_cell_value}
\alias{compute_cell_value}
\title{Given a DF of all observations (estimates) for the cell, combine them
in some way to generate a single value}
\usage{
compute_cell_value(
  num_long_df,
  unit_varname,
  unit_val,
  time_varname,
  time_val,
  combined_df
)
}
\arguments{
\item{combined_df}{}
}
\description{
For now \link{the default} just average them all, dropping any N/As, but
returning np.nan if this results in no entries
}
\details{
Combine the rows in a df of estimates into one final estimate
}
