% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_parse.R
\name{.process_data_file}
\alias{.process_data_file}
\title{Process a raw data file, transforming it into Threadsheets format}
\usage{
.process_data_file(fpath, spec, verbose = FALSE)
}
\arguments{
\item{spec}{}
}
\description{
This is the main function for this file
}
