% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_parse.R
\name{.var_in_df}
\alias{.var_in_df}
\title{Check if a variable is present in a tibble}
\usage{
.var_in_df(varname, df)
}
\arguments{
\item{varname}{The variable you want to look for in the headers of \code{df}}

\item{df}{The tibble you want to check the headers of}
}
\value{
\code{TRUE} if one of the tibble headers is either (a) \code{varname} itself,
or (b) of the form "\code{varname}@year"

df <- tribble(
~country, ~\verb{cp_membership_num@1950}, ~\verb{cp_membership_num@1970},
Austria, 10000, 5000,
Zimbabwe, 100, 50
)
.var_in_df("cp_membership_num", df)
}
\description{
We need this because of the special '@' character, which separates levels
of the MultiIndex
}
