% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_create_grid.R
\name{threader_create_grid}
\alias{threader_create_grid}
\title{Step 3: Construct the Grid of Estimates}
\usage{
threader_create_grid(
  combined_fpath = "./data/parsed/combined.rds",
  spec_fpath = "./data/demo.yaml"
)
}
\description{
Step 3: Construct the Grid of Estimates
}
