% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_parse.R
\name{threader_parse}
\alias{threader_parse}
\title{Step 1: Parse Data Files}
\usage{
threader_parse(
  data_path = "./data/",
  spec_fpath = "./data/demo.yaml",
  verbose = FALSE
)
}
\description{
Step 1: Parse Data Files
}
