% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_push_to_git.R
\name{threader_push_to_git}
\alias{threader_push_to_git}
\title{Push estimate grid to GitHub}
\usage{
threader_push_to_git(
  output_path = "./data/parsed/",
  spec_fpath = "./data/demo.yaml"
)
}
\description{
Push estimate grid to GitHub
}
