% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threader_push_to_git.R
\name{threader_push_to_git}
\alias{threader_push_to_git}
\title{Push estimate grid to GitHub}
\usage{
threader_push_to_git(combined_path = NULL, spec_fpath = NULL)
}
\arguments{
\item{combined_path}{Path to the directory containing the combined (wide-format) grid datasets you want to push to git}

\item{spec_fpath}{Filepath to the \code{.yaml}-format spec file.}
}
\value{
An environment with keys \code{num_result}, the return value from the GH API call pushing the numeric grid, and \code{info_result}, the return value from the GH API call pushing the info grid
}
\description{
Push estimate grid to GitHub
}
\examples{
# Not run, since it requires authentication (via `.env` file)
#gh_result <- threader_push_to_git()
}
